project file structure refactor
