+++ lib/auth.ts
@@ -5,7 +5,7 @@ import NextAuth from 'next-auth'
 import GitHub from 'next-auth/providers/github'
 import Google from 'next-auth/providers/google'
 import Credentials from 'next-auth/providers/credentials'
-import signInWithCredentials from './video-api/auth/signin-with-credentials'
+import signInWithCredentials from './video-api/auth/mutations/signin-with-credentials'
 import { signInFormSchema } from './validation/auth-validation'
 import { logger } from '@/lib/logger'
 // import getAccessToken from './video-api/auth/get-access-token'
