+++ lib/video-api/auth/actions/auth-actions.ts
@@ -1,11 +1,12 @@
 'use server'
 
 import type { SignUpFormData } from '@/lib/types/common.types'
-import type { User } from '@/lib/types/video-api.types'
+import type { User } from 'next-auth'
 
 import { signUpFormSchema } from '@/lib/validation/auth-validation'
 
 export async function signup(values: SignUpFormData) {
+  // Input validation with password confirmation
   const validatedFields = signUpFormSchema.safeParse(values)
 
   if (!validatedFields.success) {
@@ -17,12 +18,8 @@ export async function signup(values: SignUpFormData) {
     }
   }
 
-  // TODO: doing password encryption and hashing might be best to be done here
-
   try {
-    const signupData = checkPasswordsMatch(validatedFields.data)
-    const signupResponse = await createUser(signupData)
-    console.log('signupResponse', signupResponse)
+    const signupResponse = await createUser(validatedFields.data)
     return signupResponse
   } catch (error) {
     const errorMessage = error instanceof Error ? error.message : 'Unknown error during sign up.'
@@ -33,18 +30,6 @@ export async function signup(values: SignUpFormData) {
       status: 500,
     }
   }
-}
-
-function checkPasswordsMatch(data: SignUpFormData): Omit<SignUpFormData, 'passConf'> {
-  const { username, email, password, passConf } = data
-  if (password === passConf) {
-    return {
-      username,
-      email,
-      password,
-    }
-  }
-  throw new Error('A megadott jelszavak nem egyeznek')
 }
 
 async function createUser(formData: Omit<SignUpFormData, 'passConf'>): Promise<User> {
